api account, invitation, friends

+++ collectionsProject/Models/Friend.cs
@@ -5,7 +5,7 @@ namespace collectionsProject.Models;
 
 public partial class Friend
 {
-    public int IDfriendship { get; set; }
+    public long IDfriendship { get; set; }
 
     public string IDrequester { get; set; }
     public User Requester { get; set; }
